+++ examples/pqc/mldsa_verify_neg.c
@@ -34,7 +34,8 @@
 #include <stdio.h>
 #include <string.h>
 
-#if !defined(WOLFTPM2_NO_WRAPPER) && defined(WOLFTPM_V185)
+#if !defined(WOLFTPM2_NO_WRAPPER) && \
+    defined(WOLFTPM_MLDSA_SIGN) && defined(WOLFTPM_MLDSA_VERIFY)
 
 static void usage(void)
 {
@@ -238,12 +239,13 @@ exit:
     return rc;
 }
 
-#endif /* !WOLFTPM2_NO_WRAPPER && WOLFTPM_V185 */
+#endif /* !WOLFTPM2_NO_WRAPPER && PQC ops */
 
 #ifndef NO_MAIN_DRIVER
 int main(int argc, char *argv[])
 {
-#if !defined(WOLFTPM2_NO_WRAPPER) && defined(WOLFTPM_V185)
+#if !defined(WOLFTPM2_NO_WRAPPER) && \
+    defined(WOLFTPM_MLDSA_SIGN) && defined(WOLFTPM_MLDSA_VERIFY)
     int rc = mldsa_verify_neg_run(argc, argv);
     return (rc == 0) ? 0 : 1;
 #else